+++ cloud-bean/cloud-bean-ai/src/main/java/cc/xiaoxu/cloud/bean/ai/dto/LocalVectorDTO.java
@@ -17,5 +17,5 @@ public class LocalVectorDTO {
     private Integer index;
 
     @Schema(description = "向量")
-    private List<Double> embedding;
+    private List<Float> embedding;
 }
